Reorganize pipeline tests
