+++ tests/test_pipelines_common.py
@@ -1,0 +1,12 @@
+from diffusers.utils.testing_utils import require_torch
+
+
+@require_torch
+class PipelineTesterMixin:
+    """
+    This mixin is designed to be used with unittest.TestCase classes.
+    It provides a set of common tests for each PyTorch pipeline, e.g. saving and loading the pipeline,
+    equivalence of dict and tuple outputs, etc.
+    """
+
+    pass
